adding router and home page
