+++ src/components/App/App.js
@@ -1,28 +1,41 @@
 import React from 'react';
 import {
   ChakraProvider,
+  Flex,
   Box,
-  Text,
-  Link,
-  VStack,
-  Code,
   Grid,
   theme,
 } from '@chakra-ui/react';
-import { HomePage } from 'pages/homepage'
-import { ColorModeSwitcher } from 'components/ColorModeSwitcher';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route
+} from 'react-router-dom'
+import { HomePage } from 'pages/HomePage'
+import { Navbar } from 'components/Navbar'
+
+const routes = [
+  {
+    'component': HomePage,
+    'path': '/'
+  }
+]
 
 export const App = () => {
   return (
     <ChakraProvider theme={theme}>
-      <Box textAlign="center" fontSize="xl">
-        <Grid minH="100vh" p={3}>
-          <ColorModeSwitcher justifySelf="flex-end" />
-          <VStack spacing={8}>
-            <HomePage/>
-          </VStack>
-        </Grid>
-      </Box>
+      <Router>
+        <Flex direction="column">
+          <Navbar/>
+          <Switch>
+            {
+              routes.map(route => (
+                <Route path={route.path} component={route.component} />
+              ))
+            }
+          </Switch>
+        </Flex>
+      </Router>
     </ChakraProvider>
   );
 };
