+++ src/components/ColorModeSwitcher/ColorModeSwitcher.js
@@ -14,7 +14,6 @@ export const ColorModeSwitcher = props => {
       aria-label={`Switch to ${text} mode`}
       variant="ghost"
       color="current"
-      marginLeft="2"
       onClick={toggleColorMode}
       icon={<SwitchIcon />}
       {...props}
