+++ src/pages/HomePage/HomePage.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { 
+  Box,
+  Text,
+  VStack,
+  Grid,
+} from '@chakra-ui/react'
+
+
+export const HomePage = () => {
+  return (
+    <Box textAlign="center" fontSize="xl">
+      <Text>
+        This is the HomePage
+      </Text>
+    </Box>
+  );
+};
