--- src/pages/homepage/homepage.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Text } from '@chakra-ui/react'
-
-
-export const HomePage = () => {
-  return (
-    <Text>
-      This is the HomePage
-    </Text>
-  );
-};
