--- src/pages/homepage/index.js
@@ -1,1 +1,0 @@
-export { HomePage } from './homepage'
