+++ tools/component.tmpl
@@ -1,0 +1,7 @@
+import React from 'react';
+
+
+export const ${COMPONENT_NAME} = () => {
+  return (
+  );
+};
